backend/requirements.txt
flask
flask-cors
requests
gunicorn
python-dotenv
uvicorn
